Everyone welcome our long-lost friend, BasicPhysics!
While BasicPhysics is nearly useless, it's the only tool to use when no other physics engine is available or working.
It at least allows you to test that all the other code is running right!

This is important mainly on Linux where we don't at this time yet have a working PhysX implmentation.