Fixed many issues with llJson2List. Full SL compatibility now.
The first 4 sections of the LL JSON testcases all pass. This update
fixes llJson2List cases. The first fix is "" being passed as the JSON
text.  Also fixes non-array/non-object cases such as "test" to return
["test"].  ParseJsonNode is now JsonNode2List and returns the LSL_List.
JSON boolean are now returned instead of 0 and 1. Instead of calling
ParseJsonNode recursively from the top level, it now calls
JsonNode2ListElement, passing whether it's nested or not (not, in the
case where it's called from the top level).  JsonNode2ListElement is new
code that sometimes treats the elements as strings, when called from a
"deep" (nested) object.  It also wraps the string text "{", "}", "[" and
"]" as appropriate for objects and arrays.