Fixed #24: Purging a specific empty folder does nothing.
Turns out none of the connecting glue between the event required to purge a folder and the code that does the heavy lifting was even in existence.

I carefully made sure the code flow, method locations and names paralleled the code for emptying the trash. Smoother debugging that way.

I also found that CachedUserInfo::PurgeFolder was misnamed as it only purged the contents of the folder. I corrected that and then added a new method that re-used the old name that actually does delete the folder and its contents.

It seems the client has some cruft as it makes some interesting calls:
* When Emptying the trash, HandlePurgeInventoryDescendents() is called.  This makes sense as the trash folder should be left unmolested.
* When Purging an empty folder located in the Trash, HandlePurgeInventoryFolder() is called.  This is what used to be RemoveInventoryFolder().  This also makes sense as the folder and all of its contents, of which there are none in this case, needs to be emptied.
* When Purging a folder that has children, both HandlePurgeInventoryFolder() and HandlePurgeInventoryDescendents() are called - in that order.  This make little sense as it should have been the same case as the above: only HandlePurgeInventoryFolder(). I suspect this is cruft from back when LL may have used the viewer to keep the DB consistent...

That last shouldn't be a big issue: HandlePurgeInventoryFolder() will remove the whole shooting match and the subsequent HandlePurgeInventoryDescendents() call will find that there's nothign to purge and simply return. Though it might have a null pointer problem when it tries to use the null result of GetFolderAttributes()...  Need to fix the Empty Trash bug in order to test.