Correct a call order and rename deleteItemsInFolder.
As requested, deleteItemsInFolder has been renamed to deleteFolderContents to better reflect actual operation.

The call order of removing the folder and removing its children was backwards: in the loop it was removing the folder then removing the descendents!
While this wouldn't cause an issue in the current DB implmentation, it could if foreign keys were being maintained.