Blocked out code based on whether it is Windows specific.
Previous test was looking for Mono in every case, even though Mono can be used to compile on Windows.
However some locations that are code that will fail on Mono on all platforms I left with Mono-specific tests.

Note that _WIN32 is defined on Windows no matter the word size. MSDN says it is always defined - it is just not defined on, say, Linux!
Even better for the _WIN32 the tests would be to be done at runtime so that the kernel32 calls can be done on Windows even if the code was compiled in Linux.  I just can't seem to dig up a way to do so.

Also corrected some typonese.