Got running on native System.Data.SQLite under Linux 64bit.
An underwhelming comment for what turned out to be a lot easier than I guessed, though it took some time to hunt down a good set of instructions.
In this case, https://gnuforensics.wordpress.com/2013/09/26/how-to-compile-sqlite-system-data-sqlite-on-linux-mono-using-xbuild/ got me the answer:
I just needed to compile a new edition of System.Data.SQLite.dll as per the above, using the 2012 csproj, and then copy in my system's copy of the .so.