Lots and lots of dead code...
I attempted to only clean out code that was not apparent as to why it shoudl ever be added back in.  Code that was well commented as to why it was dead was left intact.

That said, there may be stuff I cleaned out that might be useful as a reference.

Otherwise dead code has a nasty habit of developing bitrot, which I observed in a few of these cases...