Last remaining user-visible references to OpenSim and/or Inworldz swept away.
This is the most complex change so far in this process.
For these changes the VersionInfo file was just the ticket, and that was easy, just needed some tuning.

llGetEnv did get some potentially backwards compatibility breaking changes however - see Issue #51 for the place to discuss those.

I also cleaned out a lot of local duplication and unneeded function-call passing of the version string.  In one location I may have broken ownership rules by giving Phlox direct access to VersionInfo, I'm not 100% certain.

Overall I attempted to try and keep the goals of each of these version and product name calls the same, just centralizing where they got that information from.  The main idea being that the more direct the access, the easier it is to determine what the effects are for any given change.  For example, previously it was a bear to find all the places changing the version string format would effect, now it's as simple as looking at where VersionInfo is used.