Code and Halcyon INI comments, part 1.
I've been working down all the existing Halcyon.ini stanzas trying to accurately explain them.  Not easy work on some of them!

The connections strings are especially dark as to what they are used for: there seems to be no rhyme or reasons as to why one or another is used in various sections of the code.
A very few are nicely specialized, but there doesn't seem to be much actuall difference between the uses of `storage_connection_string`, `core_connection_string`, and `legacy_inventory_source`. Each one is used in too many places for completely different purposes, so much so that I think if one were to point them at separate databases there'd be chaos...  I'm just not so certain of that that I'd actually combine them!