Fixed forced group membership db queries to be optimized if possible.
Uses the LLCV groups list if there is one. Note: it looks like more
changes since I also fixed the agentID/agentId naming inconsistency so I
could follow it through in the debugger.

The key change is to now call FastConfirmGroupMember instead of
IsAgentInGroup in the groups module (which is a db call).