Added the missing schema information for the user server xml config file.
In the process of debugging why the previous idea didn't work I found out that the User Server hardly pays any attention to the Halcyon.ini.  Needs it for some stuff, but not a lot. Adding the infrastructure to try and get that info from the ini was more intrusive than I intended to go, so I simply added the needed schema info to the xml, modified my xml, et voila, it works!

Consequently I was able to remove all the INI file reader code I'd added as it was irrelevant.

I also did some more cleanup to the default login message.