Transitions to/from local user are now done as conversion rather than forcing an update.
TryGetUserProfile is available to check if a user profile is already
cached.  Profile cache management should now be keeping the lists
consistent and using the proper lock for each.   Also updated some of
the console log messages for consistent category naming (e.g. [USER
CACHE]:) and from [PROFILE] to [USERSTORAGE] in the storage-related
functions in UserProfileManagerData.