Fixed the deadlock reported on the sandboxes (hopefully).
Split AbsolutePosition, which involves a leaf lock (m_posInfo) into a
separate function that sometimes does and sometimes does not do the
extra work involving parcel checks and updates from physics.  Only
CheckForSignificantMovement does the full version, and it's safe to call
it from that.  This is also a performance optimization to simplify the
code for all of the other calls to AbsolutePosition.