Two more Gendarme-guided corrections: DoNotDestroyStackTraceRule and one I forgot the name of.
DoNotDestroyStackTraceRule means that a throw in a catch shouldn't try to rethrow the exception instance and therefore mask the stack trace, but instead to just .

The other rule was that classes that are composed of only static members should be static.  I've only done a couple of them, but there are many more.