Implemented the RezRestoreToWorld (Restore To Last Position) viewer packet
This adds support for single-object and coalesced object Restore To Last
Position viewer commands. It differs from SL by blocking the operation
when the last position was as an attachment, and informs the user that
it was an attachment and they should use Wear or Add. (SL rezzes the
object instead, near 0,0,0 in the region.)