Don't attempt to save changes to bot attachments when bot removed
This didn't commit in the past, a bit by fluke, because
DoStoreKnownAsset skips storage for any user that it can't look up, but
this change avoids queuing the commit completely, and possible future
error cases in the persistence code that don't specifically check for
bot attachments.  Note that it's difficult to check farther down because
there is no SP and the bot is removed already or in progress.