Fixes to IsScripted initialization and maintenance
- Resolves http://jira.phoenixviewer.com/browse/FIRE-17782
- Return value from CheckIfScriptedStatusChanged wasn't used, so removed
the return value and simplified the function considerably, including
removing the special casing for the root prim.  Renamed to
RecalcScriptedStatus.
- Removed the unconditional clearing of the root prims Scripted flag in
LoadPrimsFromStorage (o.O) after the whole group was initialized,
replaced with a call to RecalcScriptedStatus to initialize the object's
IsScripted status.
- Fixed a possible NRE in AddFlag if the ParentGroup was null (as it can
be now during scene persistence backup).
- Fixed prim full updates to send the prim scripted status rather than
the object scripted status. (This is the primary fix for the JIRA above,
although the rest is also required.)
- Moved the setting of the Scripted flag to inside the loop in
CreateScriptInstances where we initialize script instances rather than
inside CreateScriptInstance where it is conditional on region settings
etc. Must be set when there are scripts present, regardless of running
state.
- Updated HandleAddRunningScript to set or clear the Scripted flag
rather than just removing it (o.O) and renamed it to
HandleChangedScripts since it's not just for adding or running scripts.