Restored the previous code but with Scripted preservation
For backwards-compatibility, I replaced the newer simple code with the
original code, but then wrapped it with new code that saves the current
state of the Scripted flag for the part, applies the old code, then
updates just that one Scripted flag with the saved value. This should
allow the previous fixes in PR#152 to be maintained.