Fixed botGetPos to not require ownership.
Made botGetPos consistent with botGetName, etc. in that any script can
get the position of a bot in the region. It is not only available to the
owner, and is not priviledged info. I think this was an unintended
copy/paste problem from botSetPos or other similar functions that do
require ownership. See
http://inworldz.com/forums/viewtopic.php?f=17&t=21338&p=194059#p194059