Fixed complete idiocy in old EmailModule
Ye olde layers of fail are so thick in this module it is difficult to
know where to begin:
- fixed the guaranteed null reference exception that was caused by
duplicating the part referencing code outside the code that checked it
for null.
- replaced that crap with something half-appropriate for the null case.
- removed the extra unneeded call to resolveNamePositionRegionName
outside the exception wrapper that was getting an exception (sigh).
- eliminated the m_ObjectRegionName function-local variable (that wasn't
a class member variable) and wasn't needed at all anyway
- eliminated the unused IsLocal method
- fixed the log4net calls that seemed to think no colon was needed after
category prefixes (e.g. "[EMAIL]:" was needed.