Fixed XMLRPC User service lookups for bots
because botCreateBot doesn't need to check with the User server if that
UUID already exists, given it just made one up.  (Call to
UserService.CacheUser is now conditional on NOT being a bot.)  Also
fixed UserProfileManager.GetUserInfo to fetch the profile indirectly
through GetUserProfile to support local profiles (presences) and temp
profiles (bots).