Fixes to position on dialog replies (button presses).
ChatModule's function FixPositionOfChatMessage was mostly hidden to the
callers who specified a location for the chat to come from.  It was a
dangerous override of the position being passed in and caused this
regression. I have removed it, callers are responsible for passing in
the desired chat source location since this code doesn't know if it's
from an avatar or a prim.  Fixes Mantis 3206.