Implemented llDetectedGrab (Mantis 2544)
- The ObjectData.GrabOffsetInitial field coming from the viewer was
always 0,0,0 and yet llDetectedGrab works in SL without it.  Looks like
a protocol design feature never implemented (or broken in the viewer and
never fixed).
- We now do what I'm assuming SL developers had to do; we store the
initial grab position on the ObjectGrab packet and use the saved one
from there on ObjectGrabUpdate.
- Mapped the unused Phlox VM.DetectVariables.Grab to the unused
parm.OffsetPos and ensure that is initialized from the saved startpos
from the grab.
- [No effective change] Removed the unused and commented-out version of
HandleObjectDeGrab to avoid confusion with the one in use.