Added ability to export the map tile when generated.
This adds the base level of functionality for making the region able to push map changes to the map server.
To set this up simply share and map the ingestion folder of the map server to the region and then point the new Halcyon.ini entry to the path and format appropriately.

At this point the region now has the tools to do the push, but still needs to be told to generate the tile.