Premature optimization is a root of many evils, but NO optimization is the cause of much hair loss.
Just moving initialization OUT of the core function and INTO to initialization operation, along with shifting some data from function-scope to object scope to gain longevity and not so much memory churn was able to increase performance of the map tile rendering signifigantly.
Specifically when TextureOnMapTile=true the speedup was over 5x, putting it nearly on par with Shaded drawing.  Shaded drawing was already pretty fast.