More CF asset handling changes
- Skip the _assetCache.HasAsset check in CacheAssetIfAppropriate since
we want to update the asset cache anyway (or it's not a LRU cache
anymore).
- Support passing in a null stream so that we can cache the full asset
on writes, *before* calling StoreAsset. Then call
CacheAssetIfAppropriate again after StoreAsset, with the stream, if one
was returned. This makes the asset available to other threads during the
asset write storage, and also in the case of an asset write (CF) timeout
or other exception.  Since the write has been issued, reads should find
the asset.
- statGetComplete has been renamed to statGetFetches to help indicate
it's only the counter of the actual fetch calls, not all reads.
- the statDupUpdate case has been eliminated since the cache update is
no longer conditional on HasAsset