Remote Admin: Added some sanity, fixed message bug.
Fixed bug where the 1 minute alerts were not being sent.
My fix uses a timer so that the thread was allowed to keep moving, resulting in the XMLRPC response being immediate instead delayed until the countdown stops.
It could have been done in a less invasive manner, but I didn't like the delayed return to the call.

Also fixed a bug where you could start multiple shutdowns or restarts in any combination.

The implentation of both of these fixes was done in a way that assumes that the server is only running a single region.  This means that this new code is another location where multiregion is broken.
I'm thinking it's time to clean up the codebase by removing that multiregion cruft.  Leave the external APIs, but internally excise the concept - we are either going to have to support the idea of multiregion or remove it, this half state is bad.