Added alternate authentication for remote access.
Under Mono there's no way to check the username+password against the user, mainly because there's no guarantee that passwords are even used.

I realized that this system was acting more as an auth token than a user login. So I simplified the concept to a single passcode.
However, their's no guarantee that the configuration files or the system's memory will be kept secure.  So I decided to keep the code around only as a salted SHA256 hash.
I didn't go to the extreme and implement key stretching, but that could be added if there's enough worry.

Under .NET on Windows the old active user authentication process is still used if the hash is blank or not set.  This provides backwards compatability with pre-existing setups.