More effecient normal calc, finer resolution timings.
The calculations of the face normals was quick to code, but not fast to execute: lots of temporary struct instances and wasted calculations.  Now drastically simplified.

Added some finer resolution to the metrics to find the slowest code - everything's getting about as effecitent as I think it can be.

Removed some dead code - never going to be used IMHO.