Login bans and other user errors were treated as region down, causing problems.
This meant if a banned user tried to log in a lot, they could deny
direct logins to that region by ALL users, again and again. See
http://inworldz.com/forums/viewtopic.php?p=208345#p208345 Fixed any good
response from the region (including user fails) to not be interpreted as
a region down failure blocking valid users after 10 errors in 5 minutes.
Also renamed the booleans to be clearer, and fixed the return code to
use both.