Support for preserving AvatarMovesWithPart across region crossings.
Note: AvatarMovesWithPart == true is legacy mode, I had to invert that
to:
AvatarAsAPrim = !AvatarMovesWithPart;
in order to default to legacy (false) on the agent protocol.  This
should give us old->new compatibility during the rollout, but not
new->old, although there probably won't be any objects expecting the new
behavior until after the rollout completes (except on the Sandboxes).