Fixed crossings broken by avatar-as-a-prim change, many other fixes.
- fixed crossings broken by seated avatar data changes
- MakeChildAgent now saves the former position as the recovery location
for reentry failures rather than 0,0,0. Also sets the avatar parent
position to 0,0,0 within the lock when clearing the parent prim.
- CompleteMovement does not override the position if there is a parent
prim, the SP is in AaaP mode, or if the requested sit target ID is zero.
- CompleteMovement's m_flying renamed to flying since it is not a member
variable.
- SP.Reset no longer sets the AbsolutePosition to 128, 128, 70 since
that had all kinds of unintended side-effects.
- SP.ChildAgentDataUpdate renamed to ChildAgentPositionUpdate since it
updated from a AgentPosition (only), rather than an AgentData.
- SP.ChildAgentPositionUpdate now finishes updating the avatar position
*before* calling the object culling code to
CheckForDistantEntitiesToShow
- SP.CopyToForRootAgent now includes the SatOnGroup, SatOnPrim, and
SatOnPrimOffset members.
- SP.CopyToForRootAgent and SP.CopyFrom are now much more careful to
grab consistent data (inside the PosInfo lock), and reorder some of the
assignments for maximum safety and consistency.
- SendAvatarState's SendChildAgentUpdate2 only updates (overrides) the
position if there is no RideOnPart and no longer ever updates the SatOn
variables.