Fixed duplicate CHANGED_LINK messages
rearranged AvatarPartsCollection to return a boolean as to whether an
Add or Remove made changes, renamed the functions to be clearer, don't
repeat the add/remove if it has already been done, and don't signal the
event times the number of prims to every prim.  Moved the event sending
to the SOG since we need to do it all once per object.