Experimental changes to crossings, dest region confirms crossing, not viewer.
Instead of waiting for the CompleteMovement packet from the viewer, the
region confirms the crossing immediately if allowed and it has upgraded
to root agent.  There are better ways to do it (like remove the
confirmation and use the agent2 put response OK instead), but this is a
good test with smaller changes.