Make sure the agent is only released once
There's a race between a fast viewer response and the server response.
Server will almost always come in first, but on a local test server it
might not, or perhaps if the server was under load. I saw it a couple of
times during my local testing.  Solution: grab the non-null
m_callbackURI in a lock and clear when grabbed.