GetLinkParts to return list of SOPs and now optionally include SPs
Returns a single list of both.  Also GetLinkPrimsOnly is the prim-only
compatible version so we can see explicitly where avatars are excluded.
This commit doesn't change much but the structure.  Next one will
include more uses and replace separate calls for prims and avatars with
a single list of both.