Fixed problem with sit target sitter being cleared on backup
Calls to SetParent from the backup code (CopyPart) operate on a
duplicate of the in-world SOG/SOP, which have a matching UUID to an
object instance in-world. When SetParent calls SetSitTarget, it finds
the in-world copy by UUID and updates the dictionary based on the
sitInfo of the DUPLICATE copy, which doesn't have a sitter on it.  (The
avatar is only sitting on one object at a time.)  When it found the
UUID, it replaced the sit target info with the available sit target,
allowing a second avatar on the same sit target, even with others
available.