Replaced some GetSceneObjectPart calls with a new GetSitTargetPart.
Replaced GetSceneObjectPart(m_requestedSitTargetUUID) calls with a new
GetSitTargetPart() for better self-documenting code. In some cases you
want the seated avatar's visual parent (e.g. for updates), but in some
cases (LSL scripting) you often want the original sat-upon prim (e.g.
for runtime permissions checks for scripts).  This is an attempt to
differentiate more obviously in the code between the two different usage
patterns.  The most obvious place where this would mattered was in
LSLSystemAPI.cs GetImplicitPermissions() which could cause implicit
permissions to apply only to scripts in the root prim, however in that
case the prim owner is checked, which should be the same for the child
part and the root prim.  The other place was in llPushObject where the
position would have been incorrect for the seated avatar, but that
wouldn't matter in llPushObject for a seated avatar.