Restored the full HasConnectionsEstablishing() check
rather than only calling HasEstablishedConnection(destRegion).

On further discussions with Dave, and after considering alternatives
(destination region sending enable/disable for whole chain of
crossings), we cannot allow an avatar to initiate a new crossing while
connections are being established. The chronological order must be
preserved, especially given the possibility of reentry into a region,
and we cannot ensure that one an avatar leaves the region.  It's a shame
that we need to block and wait for more than just the destination
region, but it is what it is, the alternatives are worse.