Fixed shallow copy (reference copying) in SOG.Copy
Several SOG members are references to other objects and the
MemberwiseClone() copies the _reference_. So changes to things like sit
target lists, or LSL scripted targets, bounding box, etc in one copy of
a SOG affected all copies of that SOG that were shift-copied. This
resulted in Mike's report of sit targets being cleared in the original
_after_ a shift-copy, and is probably responsible for some of the LSL
target-related bugs, as well as other object "quantum entanglement"
reports.