Minor: keep the SOG name in sync with root part
Changes to the SOG.Name also update the object name to keep it in sync
with the root part name. This mostly only affects debugging and other
symbolic dumps since the Name getter override in SOG pulls the name from
the root part.  But this keeps the objects consistent just in case.