Fixed NewAgentInventoryRequest CAP to send an inventory item update on upload
This fixes the item permissions being out of sync with the default
permissions after an upload (until the first relog/refresh). This is
really a viewer bug since it is notifying the server of the permissions
to use, but we can ensure all viewers have the correct permissions,
without a relog, by sending the update after item creation.