Implemented llReturnObjectsByOwner and llReturnObjectsByID
Also clarified and simplified cases where a group was checked and NOT
being a member was treated as 0 (member in Everyone role) by using a
nullable type for group powers.  Factored caching of group powers, and
otherwise also favor use of LLCV client cached powers if available when
not cached.