Lots of tweaks, must NOT use Permissions.CanUseObjectReturn
Those are user-session / viewer permissions checks which do not match
the scripted checks (almost at all). Also moved canUseReturnPermission
check to the top of the functions since we don't need to iterate at all
if it fails.  Lots of little fixes with the return codes and subtle
fixes to check semantics.