Fix script engine crash seen on SB5, SB6 and Plus SB.
Avoid a race condition where the viewer might send what results in a
second _Replace(itemID, null) for a delete which is a *duplicate*.  This
can happen if the user deletes the TaskInventoryItem a *second* time
before the viewer updates the Contents to remove it. This change just
handle the itemID no longer being present in the items list. Also
replaced the magic numbers with InventoryType constants.
See
https://inworldz.zendesk.com/hc/en-us/community/posts/115001158408-Sandboxes-updated-to-0-9-32-test-version