Fixed FetchInventoryDescendentsRequest to fetch folder contents including links without following links
Just retrieving the listing should not include the actual items linked
to, especially in the case of folder links. In both cases, this probably
results in both the link AND the original being present in the viewer's
folder content, which could be particularly nasty for folders since
their parent may not yet be loaded at the viewer end when we send the
folder down. This change just sends links as links when filling a folder
list, and does not attempt to follow them.