Restore the previous level of checking on exiting a region
- Added a new CanExitRegion which differentiates from IsFullyInRegion
only for the purposes of the avatar exit checks. This restores the
previous checks for avatar region exist only, while preserving the
changes for avatar/culling updates etc.
- Also renamed the FullyInRegion flag to what it is, ParcelInfoSent.
FullyInRegion is more of a subjective test by the caller, which includes
that flag but others as well. (No change to generated code for this
part, just a variable rename.)