Fixed exception saving scripts under updated viewers.
Newer viewers are using the sim CAP now, which passes a bool, but LLSD
doesn't actually use a real bool, it's an int under the covers, so we
need to convert or we get a C# type exception. This matches the fix in
OpenSim for the same lack of support. This should actually be fixed at
the OSD level but until that happens, this change should avoid the
exceptions for all LLSD messages that include bools.