Bandwidth is seriously underused (capped at 250KB/s) so let's allow more for neighbors
This might change back to 30% with more viewer tuning and higher
bandwidth settings coming. Also became a problem with neighbors range of
2 since that increased neighbors from 8 to 24.  1/24th of 30% is too
slow of a trickle, especially when we use 2/3rds of the bandwidth of
other grids and we also use it to limit textures via Aperture (nobody
else does).