Mitigate against a late CompleteAgent retry which totally screws things up
This request is handled with
PacketHandlingDisposition.HandleSynchronized however it has an async
section at the end, and no protection from queuing up the duplicate
request for serial processing once the first call is complete. We're
tracking AgentInRegion status here so just avoid resetting it at the
start, and instead, use it to recognize the duplicate.