Replaced no longer functional __MonoCS__ flag with custom flag
`__MonoCS__` was always considered a hack, and with Mono's conversion to msbuild it's no longer defined.  Intsead of leaving the old semantic, I've instead declared a new flag, `_MONO_CLI_FLAG_`, that says where's it's from and what it means.

The AppVeyor build script likewise declares it, and anyone compiling with Mono should likewise declare it until such time as this Windows-specific code can be excised.