Add ability to Appveyor build artifacts to GitHub Releases on tag
Release procedure will be to commit a change that changes the VersionIfo.tt file to the new version being released, push it, tag it, push the tag.  When AppVeyor finishes and uploads to GitHub a draft release will then be ready for release notes and a note that the binary release is "as is" and built on an uncontrolled 3rd party server and shipped with libraries built on other uncontrolled 3rd party servers - don't blame us if it steals your moneys and nukes your server; compile yer own.